Added a tabpane, tab for plotting
Added a QTabWidget to be able to switch between table and plot view.
Added a plotter which plots up the cumulative sum per day. Added
signaling to DbHandler so that reloads must not be explicitly stated in
code. UI elements depending on the database should simply be connected
to that signal.

Signed-off-by: Max Franke <max@mumintroll.org>