Fixed a bug with the zooming in plots
Formerly, when zooming past a certain zoom level, the plot graph
would overshoot the plotting area and not be able to be scrolled
to.
Fixed this by using the dticonverter for determining the rectangle
for the background as well. Also introduced a dedicated rectangle
for the total background, including the axes.