Implemented backup functionality
The script backup.sh should be installed to the user config location or
someplace on installation for this to work. The script should take one
argument, which is the database location. This way, we need not hardcode
it.
There is still some work to do here later on, but for now, this should
be fine.

Re #28