Made choice of x terminal emulator more robust
Previously, the terminal emulator was hard coded. When I changed to a
new system where rxvt-unicode was not yet installed, I noticed that this
was not a good idea.
The program will now try to find a terminal emulator from a list. If it
finds none, it will use /bin/false. This is part of the GNU coreutils.
If even that is not installed, the effect will still be the same --
nothing.